refactor code into multiple files
