+++ pkg/client/sdk_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package provider
+package client
 
 import (
 	"errors"
