+++ pkg/provider/list_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/gardener/machine-controller-manager/pkg/util/provider/machinecodes/status"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/stackitcloud/machine-controller-manager-provider-stackit/pkg/client"
 	api "github.com/stackitcloud/machine-controller-manager-provider-stackit/pkg/provider/apis"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -72,10 +73,10 @@ var _ = Describe("ListMachines", func() {
 
 	Context("with valid inputs", func() {
 		It("should list machines filtered by MachineClass label", func() {
-			mockClient.listServersFunc = func(_ context.Context, _, _ string, selector map[string]string) ([]*Server, error) {
+			mockClient.listServersFunc = func(_ context.Context, _, _ string, selector map[string]string) ([]*client.Server, error) {
 				Expect(selector["mcm.gardener.cloud/machineclass"]).To(Equal("test-machine-class"))
 
-				return []*Server{
+				return []*client.Server{
 					{
 						ID:   "server-1",
 						Name: "machine-1",
@@ -106,8 +107,8 @@ var _ = Describe("ListMachines", func() {
 		})
 
 		It("should return empty list when no servers match", func() {
-			mockClient.listServersFunc = func(_ context.Context, _, _ string, _ map[string]string) ([]*Server, error) {
-				return []*Server{}, nil
+			mockClient.listServersFunc = func(_ context.Context, _, _ string, _ map[string]string) ([]*client.Server, error) {
+				return []*client.Server{}, nil
 			}
 
 			resp, err := provider.ListMachines(ctx, req)
@@ -118,8 +119,8 @@ var _ = Describe("ListMachines", func() {
 		})
 
 		It("should return empty list when no servers exist", func() {
-			mockClient.listServersFunc = func(_ context.Context, _, _ string, _ map[string]string) ([]*Server, error) {
-				return []*Server{}, nil
+			mockClient.listServersFunc = func(_ context.Context, _, _ string, _ map[string]string) ([]*client.Server, error) {
+				return []*client.Server{}, nil
 			}
 
 			resp, err := provider.ListMachines(ctx, req)
@@ -132,7 +133,7 @@ var _ = Describe("ListMachines", func() {
 
 	Context("when STACKIT API fails", func() {
 		It("should return Internal error on API failure", func() {
-			mockClient.listServersFunc = func(_ context.Context, _, _ string, _ map[string]string) ([]*Server, error) {
+			mockClient.listServersFunc = func(_ context.Context, _, _ string, _ map[string]string) ([]*client.Server, error) {
 				return nil, fmt.Errorf("API connection failed")
 			}
